re-ordered sweep operations
re-organized the sweep operations so that the column 1 and column 2
photoresistor inputs get read as close together as possible. (may be
able to do this via port manupulation much faster but i dont know how
right now).  The old way had a for loop that delayed the reading by at
least the cycle time (2ms).